+++ collector/kernel/kernel_collector.cc
@@ -100,7 +100,7 @@ KernelCollector::KernelCollector(
     std::optional<AuthzFetcher> &authz_fetcher,
     bool enable_http_metrics,
     bool enable_userland_tcp,
-    double socket_stats_interval_sec,
+    u64 socket_stats_interval_sec,
     CgroupHandler::CgroupSettings cgroup_settings,
     ProcessHandler::CpuMemIoSettings const *cpu_mem_io_settings,
     std::string const &bpf_dump_file,
